reorganized src files a bit so hopefully it's a little more sensible. also updated some things, e.g. switched to evt.code instead of evt.which, hash-name methods (#) instead of underscore for private class methods
