+++ .babelrc
@@ -1,6 +1,6 @@
 {
   "presets": [
-    "@babel/preset-env", 
+    ["@babel/preset-env", { targets: "defaults" }],
     "@babel/preset-react",
   ],
   "plugins": ["@babel/plugin-transform-runtime"]
