+++ jest.config.json
@@ -3,6 +3,6 @@
   "roots": ["tests"],
   "testRegex": "((\\.|/*.)(spec))\\.js?$",
   "moduleNameMapper": {
-	"^.+\\.css$": "<rootDir>/tests/cssStub.js"
+    "^.+\\.css$": "<rootDir>/tests/cssStub.js"
   }
 }
