+++ src/FilterManager.js
@@ -1,27 +1,27 @@
-import { Saturation } from '../filters/saturation.js';
-import { Grayscale } from '../filters/grayscale.js';
-import { AreaColor } from '../filters/areacolor.js';
-import { EdgeDetection } from '../filters/edgedetection.js';
-import { Invert } from '../filters/invert.js';
-import { Mosaic } from '../filters/mosaic.js';
-import { Blur } from '../filters/blur.js';
-//import { SimpleBlur } from '../filters/simple_blur.js';
-import { TargetedBlur } from '../filters/targetedBlur.js';
-import { Outline } from '../filters/outline.js';
-import { Voronoi } from '../filters/voronoi.js';
-import { Fisheye } from '../filters/fisheye.js';
-import { HorizontalShift } from '../filters/shift.js';
-import { CRT } from '../filters/crt.js';
-import { ChannelShift } from '../filters/channel_shift.js';
-import { Dots } from '../filters/dots.js';
-import { Dots2 } from '../filters/dots2.js';
-import { Dots3 } from '../filters/dots3.js';
-import { Lines } from '../filters/lines.js';
-import { Thinning } from '../filters/thinning.js';
-//import { Solidify } from '../filters/solidify.js';
-import { OilPainting } from '../filters/oilpainting.js';
-import { Painted } from '../filters/painted.js';
-import { Wavy } from '../filters/wavy.js';
+import { Saturation } from './filters/saturation.js';
+import { Grayscale } from './filters/grayscale.js';
+import { AreaColor } from './filters/areacolor.js';
+import { EdgeDetection } from './filters/edgedetection.js';
+import { Invert } from './filters/invert.js';
+import { Mosaic } from './filters/mosaic.js';
+import { Blur } from './filters/blur.js';
+//import { SimpleBlur } from './filters/simple_blur.js';
+import { TargetedBlur } from './filters/targetedBlur.js';
+import { Outline } from './filters/outline.js';
+import { Voronoi } from './filters/voronoi.js';
+import { Fisheye } from './filters/fisheye.js';
+import { HorizontalShift } from './filters/shift.js';
+import { CRT } from './filters/crt.js';
+import { ChannelShift } from './filters/channel_shift.js';
+import { Dots } from './filters/dots.js';
+import { Dots2 } from './filters/dots2.js';
+import { Dots3 } from './filters/dots3.js';
+import { Lines } from './filters/lines.js';
+import { Thinning } from './filters/thinning.js';
+//import { Solidify } from './filters/solidify.js';
+import { OilPainting } from './filters/oilpainting.js';
+import { Painted } from './filters/painted.js';
+import { Wavy } from './filters/wavy.js';
 
 class FilterManager {
   constructor(animationProject, brush){
